Add files via upload
Updates in preparation for student usage off the rollers. Refined maximum speed to amount to about 0.45 m/s on the actual robot. Also refined acceleration and deceleration limits to prevent aggressive speed changes that might tip Merry. Adds back in the -1 mode which sets all gains to 0. The ros / manual switch now takes advantage of this gain adjustment to make it possible to push Jinx around even when the start_jinx launchfile is running.